feat: Classify videos by category and provide suggestions based on friends and favorite genres

+++ learnvocabulary/src/main/java/com/phatpl/learnvocabulary/models/graph/NVideo.java
@@ -22,5 +22,5 @@ public class NVideo {
 
     @Property("video_id")
     @NotNull
-    private Integer videoId;
+    private Long videoId;
 }
